Add and activate jekyll-srcset2 plugin
This plugin is used to generate several different sizes of each
image and serve the one most appropriate for both the size of the
image on the page and the pixel density of the device. This commit
both updates the Gemfile to install the plugin and applies the
changes in the template files. Note that this has installed
jekyll-srcset2 rather than jekyll-srcset due to a bug in the
main plugin that doesn't generate the correct image sizes.